clarify installation and usage instructions
It took some piecing together of various forum comments in order to get this working for me. The instructions in the README.md didn't work for me at all (with Heroku), but the ones I have added worked perfectly. Additionally, I'm happy to take any extra information or clarification in to consideration in order to help clarify this documentation as this is a pretty great package to have around.